requirements.txt
gitpython
pyyaml
openai
backoff
